+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+## 3.3.2
+
+* Fixing [#57](https://github.com/krasimir/navigo/issues/57) - making sure that we keep the last resolved url when we have the `notFound` handler resolves.
+
 ## 3.3.1
 
-* Fixing #56 - When a `/` is given as a path it is considered as a default route handler.
+* Fixing [#56](https://github.com/krasimir/navigo/issues/56) - When a `/` is given as a path it is considered as a default route handler.
 
 ## 3.3.0